Пересоберёт типы
