+++ src/types/authorized-user.ts
@@ -1,0 +1,10 @@
+import { User } from './user.ts';
+
+type Authorization = {
+  'email': string;
+  'token': string;
+};
+
+type AuthorizedUser = Authorization | User;
+
+export type { AuthorizedUser };
